refactor: move createDashboardViaApi to dashCreation.js
- Move createDashboardViaApi from apiCleanup.js to dashCreation.js
  (dashboard helper is the correct location, not cleanup file)
- createReportViaApi now handles MetaDashboard format when fetching
  dashboards (unwraps nested v8/v7 version keys)
- Test imports createDashboardViaApi from dashCreation.js
- Add cleanupDashboards call to cleanup test
